ProjectRename: Rename library to py_to_proto
