+++ .isort.cfg
@@ -7,4 +7,4 @@ import_heading_thirdparty=Third Party
 import_heading_firstparty=First Party
 import_heading_localfolder=Local
 known_firstparty=alog
-known_localfolder=jtd_to_proto,tests
+known_localfolder=py_to_proto,tests
