+++ py_to_proto/compat_annotated.py
